Refactor manifest module into submodules
